# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/categories": {
      "post": {
        "description": "Creates a new transaction category",
        "operationId": "CategoriesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCategoryDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Category created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation error"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Create category",
        "tags": ["categories"]
      },
      "get": {
        "description": "Retrieves all categories for the authenticated user, optionally filtered by transaction type",
        "operationId": "CategoriesController_findAll",
        "parameters": [
          {
            "name": "type",
            "required": false,
            "in": "query",
            "description": "Filter by transaction type",
            "schema": {
              "$ref": "#/components/schemas/TransactionType"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Categories retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CategoryResponseDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Get all categories",
        "tags": ["categories"]
      }
    },
    "/categories/{id}": {
      "get": {
        "description": "Retrieves a specific category by its ID",
        "operationId": "CategoriesController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Category UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Category retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Category not found"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Get category by ID",
        "tags": ["categories"]
      },
      "put": {
        "description": "Updates an existing category",
        "operationId": "CategoriesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Category UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCategoryDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Category updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Category not found"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Update category",
        "tags": ["categories"]
      },
      "delete": {
        "description": "Soft deletes a category",
        "operationId": "CategoriesController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Category UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Category deleted successfully"
          },
          "404": {
            "description": "Category not found"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Delete category",
        "tags": ["categories"]
      }
    },
...
  },
  "info": {
    "title": "Track My Money API",
    "description": "API documentation for Track My Money application",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Enter JWT token",
        "in": "header"
      },
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "refreshToken"
      }
    },
    "schemas": {
...
      "TransactionType": {
        "type": "string",
        "enum": ["EXPENSE", "INCOME"],
        "description": "Transaction type"
      },
      "CreateCategoryDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Category name",
            "example": "Groceries",
            "minLength": 1,
            "maxLength": 100
          },
          "type": {
            "description": "Transaction type",
            "example": "EXPENSE",
            "allOf": [
              {
                "$ref": "#/components/schemas/TransactionType"
              }
            ]
          },
          "parentCategoryId": {
            "type": "object",
            "description": "Parent category ID (for subcategories)",
            "format": "uuid",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          }
        },
        "required": ["name", "type"]
      },
      "CategoryResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Category ID",
            "format": "uuid",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "name": {
            "type": "string",
            "description": "Category name",
            "example": "Groceries"
          },
          "type": {
            "description": "Transaction type",
            "example": "EXPENSE",
            "allOf": [
              {
                "$ref": "#/components/schemas/TransactionType"
              }
            ]
          },
          "parentCategoryId": {
            "type": "object",
            "description": "Parent category ID",
            "format": "uuid",
            "example": "123e4567-e89b-12d3-a456-426614174000",
            "nullable": true
          },
          "parentCategory": {
            "description": "Parent category",
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/CategoryResponseDto"
              }
            ]
          },
          "subcategories": {
            "description": "Subcategories",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CategoryResponseDto"
            }
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "Creation date",
            "example": "2024-01-01T00:00:00.000Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "description": "Last update date",
            "example": "2024-01-01T00:00:00.000Z"
          }
        },
        "required": [
          "id",
          "name",
          "type",
          "parentCategoryId",
          "parentCategory",
          "subcategories",
          "createdAt",
          "updatedAt"
        ]
      },
      "UpdateCategoryDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Category name",
            "example": "Groceries",
            "minLength": 1,
            "maxLength": 100
          },
          "type": {
            "description": "Transaction type",
            "example": "EXPENSE",
            "allOf": [
              {
                "$ref": "#/components/schemas/TransactionType"
              }
            ]
          },
          "parentCategoryId": {
            "type": "object",
            "description": "Parent category ID (for subcategories)",
            "format": "uuid",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          }
        },
        "required": ["name", "type"]
      },
...
    }
  }
}